refactor(ansi): migrate from experimental ansi to stable github.com/charmbracelet/x/ansi
- replaced all imports of github.com/charmbracelet/x/exp/term/ansi with github.com/charmbracelet/x/ansi
- updated ReadFileContent in utils/file_utils.go to use stable ansi
- ran go mod tidy to remove old dependency and update go.sum
