+++ src/internal/common/load_config.go
@@ -10,7 +10,7 @@ import (
 	"reflect"
 	"runtime"
 
-	"github.com/charmbracelet/x/exp/term/ansi"
+	"github.com/charmbracelet/x/ansi"
 	"github.com/pelletier/go-toml/v2"
 
 	"github.com/yorukot/superfile/src/config/icon"
@@ -103,7 +103,6 @@ func ValidateConfig(c *ConfigType) error {
 // with the default values and modify the hotkeys if the FixHotkeys flag is on.
 func LoadHotkeysFile() {
 	err := utils.LoadTomlFile(variable.HotkeysFile, HotkeysTomlString, &Hotkeys, variable.FixHotkeys)
-
 	if err != nil {
 		userMsg := fmt.Sprintf("%s%s", LipglossError, err.Error())
 
@@ -194,12 +193,12 @@ func LoadAllDefaultConfig(content embed.FS) {
 	}
 
 	// Prevent failure for first time app run by making sure parent directories exists
-	if err = os.MkdirAll(filepath.Dir(variable.ThemeFileVersion), 0755); err != nil {
+	if err = os.MkdirAll(filepath.Dir(variable.ThemeFileVersion), 0o755); err != nil {
 		slog.Error("Error creating theme file parent directory", "error", err)
 		return
 	}
 
-	err = os.WriteFile(variable.ThemeFileVersion, []byte(variable.CurrentVersion), 0644)
+	err = os.WriteFile(variable.ThemeFileVersion, []byte(variable.CurrentVersion), 0o644)
 	if err != nil {
 		slog.Error("Error writing theme file version", "error", err)
 	}
@@ -230,7 +229,7 @@ func WriteThemeFiles(content embed.FS) error {
 	_, err := os.Stat(variable.ThemeFolder)
 
 	if os.IsNotExist(err) {
-		if err = os.MkdirAll(variable.ThemeFolder, 0755); err != nil {
+		if err = os.MkdirAll(variable.ThemeFolder, 0o755); err != nil {
 			slog.Error("Error creating theme directory", "error", err)
 			return err
 		}
