+++ src/internal/ui/rendering/border.go
@@ -3,7 +3,7 @@ package rendering
 import (
 	"strings"
 
-	"github.com/charmbracelet/x/exp/term/ansi"
+	"github.com/charmbracelet/x/ansi"
 
 	"github.com/charmbracelet/lipgloss"
 )
